Made paths static and defined them on first-use
The URI in particular doesn’t get used much at all and probably
shouldn’t be defined as soon as the class is spun up. The paths also
shouldn’t change from instance to instance, so they may as well be
static.

This needs some more thought and probably another set of eyes.