Stopped using wp_get_theme raw
Even though this is cached, the object that’s spun up every time this
function is called is pretty complex.

I can’t think of any situations in our library when we’d need to
continuously spin up this object, so grabbing it once per theme/parent
should be sufficient and perform a bit better.